refactor filter (no tests)
