+++ src/main/webapp/app/localci/build-queue/build-queue.component.scss
@@ -2,18 +2,6 @@
     word-break: break-all; /* This will break the text at any character */
 }
 
-.finished-build-jobs-filter-border-bottom {
-    border-bottom: 2px solid var(--overview-blue-border-color);
-}
-
-.finished-build-jobs-filter-background-accent {
-    background: var(--user-management-background-color);
-}
-
-.form-item-prepend {
-    width: 13%;
-}
-
 .finish-jobs-column-tiny {
     max-width: 20px;
 }
