+++ src/main/webapp/app/localci/build-queue/finished-builds-filter-modal/finished-builds-filter-modal.component.html
@@ -1,0 +1,138 @@
+<div class="modal-content">
+    <div class="modal-header">
+        <h5 class="modal-title">
+            <span jhiTranslate="artemisApp.buildQueue.filter.title"></span>
+        </h5>
+        <button type="button" class="btn-close" aria-label="Close" (click)="cancel()"></button>
+    </div>
+    <div class="modal-body">
+        <div class="container-fluid">
+            <div class="row finished-build-jobs-filter-border-bottom">
+                <div class="col d-flex flex-row pb-1">
+                    <h5 class="my-0">
+                        <span jhiTranslate="artemisApp.buildQueue.filter.buildStatus.title"></span>
+                    </h5>
+                </div>
+            </div>
+            <div class="row finished-build-jobs-filter-background-accent">
+                <div class="col container px-3">
+                    <ul class="text-nowrap list-unstyled row my-0 py-2">
+                        @for (status of this.buildStatusFilterValues; track status) {
+                            <li class="col">
+                                <label>
+                                    <input
+                                        class="form-check-input"
+                                        (change)="this.toggleBuildStatusFilter(status)"
+                                        [checked]="this.finishedBuildJobFilter.status === status"
+                                        type="radio"
+                                    />
+                                    <span class="ps-1" jhiTranslate="artemisApp.buildQueue.filter.buildStatus.{{ status }}"></span>
+                                </label>
+                            </li>
+                        }
+                        <li class="col-3">
+                            <input class="form-check-input" type="radio" (click)="this.toggleBuildStatusFilter()" [checked]="this.finishedBuildJobFilter.status === undefined" />
+                            <label class="form-check-label ps-1">
+                                <span jhiTranslate="artemisApp.buildQueue.filter.none"></span>
+                            </label>
+                        </li>
+                    </ul>
+                </div>
+            </div>
+
+            @if (buildAgentFilterable) {
+                <div class="row finished-build-jobs-filter-border-bottom">
+                    <div class="col d-flex flex-row pb-1">
+                        <h5 class="my-0">
+                            <span jhiTranslate="artemisApp.buildQueue.filter.buildAgentAddress"></span>
+                        </h5>
+                    </div>
+                </div>
+                <div class="row finished-build-jobs-filter-background-accent">
+                    <div class="col container px-3 my-0 py-2">
+                        <input
+                            type="text"
+                            class="form-control"
+                            [placeholder]="'artemisApp.buildQueue.filter.buildAgentAddress' | artemisTranslate"
+                            [(ngModel)]="finishedBuildJobFilter.buildAgentAddress"
+                            [ngbTypeahead]="typeaheadSearch"
+                            (focus)="focus$.next($any($event).target.value)"
+                            (click)="clickEvents($event, addressTypeahead)"
+                            (change)="filterBuildAgentAddressChanged()"
+                            #addressTypeahead="ngbTypeahead"
+                        />
+                    </div>
+                </div>
+            }
+            <div class="row finished-build-jobs-filter-border-bottom">
+                <div class="col d-flex flex-row pb-1">
+                    <h5 class="my-0">
+                        <span jhiTranslate="artemisApp.buildQueue.filter.buildSubmissionDate.title"></span>
+                    </h5>
+                </div>
+            </div>
+            <div class="row finished-build-jobs-filter-background-accent">
+                <div class="col container px-3 my-0 py-2">
+                    <div class="input-group">
+                        <div class="input-group-prepend form-item-prepend">
+                            <span class="input-group-text" jhiTranslate="artemisApp.buildQueue.filter.buildSubmissionDate.from"></span>
+                        </div>
+                        <jhi-date-time-picker
+                            id="field_startDate"
+                            [shouldDisplayTimeZoneWarning]="false"
+                            [labelName]="''"
+                            [(ngModel)]="finishedBuildJobFilter.buildSubmissionDateFilterFrom"
+                            [error]="!finishedBuildJobFilter.areDatesValid"
+                            (valueChange)="filterDateChanged()"
+                        />
+                    </div>
+                    <div class="input-group">
+                        <div class="input-group-append form-item-prepend">
+                            <span class="input-group-text" jhiTranslate="artemisApp.buildQueue.filter.buildSubmissionDate.to"></span>
+                        </div>
+                        <jhi-date-time-picker
+                            id="field_endDate"
+                            [shouldDisplayTimeZoneWarning]="false"
+                            [labelName]="''"
+                            [(ngModel)]="finishedBuildJobFilter.buildSubmissionDateFilterTo"
+                            [error]="!finishedBuildJobFilter.areDatesValid"
+                            (valueChange)="filterDateChanged()"
+                        />
+                        @if (!finishedBuildJobFilter.areDatesValid) {
+                            <span class="invalid-feedback" jhiTranslate="artemisApp.buildQueue.filter.buildSubmissionDate.invalidDate"></span>
+                        }
+                    </div>
+                </div>
+            </div>
+
+            <div class="row finished-build-jobs-filter-border-bottom">
+                <div class="col d-flex flex-row pb-1">
+                    <h5 class="my-0">
+                        <span jhiTranslate="artemisApp.buildQueue.filter.buildDuration.title"></span>
+                    </h5>
+                </div>
+            </div>
+            <div class="row finished-build-jobs-filter-background-accent">
+                <div class="col container px-3 my-0 py-2">
+                    <div class="input-group">
+                        <span class="input-group-text" jhiTranslate="artemisApp.buildQueue.filter.buildDuration.lowerBound"></span>
+                        <input type="number" class="form-control" [(ngModel)]="finishedBuildJobFilter.buildDurationFilterLowerBound" (change)="filterDurationChanged()" />
+                        <span class="input-group-text" jhiTranslate="artemisApp.buildQueue.filter.buildDuration.upperBound"></span>
+                        <input type="number" class="form-control" [(ngModel)]="finishedBuildJobFilter.buildDurationFilterUpperBound" (change)="filterDurationChanged()" />
+                        @if (!finishedBuildJobFilter.areDurationFiltersValid) {
+                            <span class="invalid-feedback" jhiTranslate="artemisApp.buildQueue.filter.buildDuration.invalidState"></span>
+                        }
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="modal-footer">
+        <button type="button" class="btn btn-secondary" (click)="cancel()">
+            <span jhiTranslate="artemisApp.buildQueue.filter.close"></span>
+        </button>
+        <button class="btn btn-primary" (click)="confirm()">
+            <span jhiTranslate="artemisApp.buildQueue.filter.apply"></span>
+        </button>
+    </div>
+</div>
