+++ src/main/webapp/app/localci/build-queue/finished-builds-filter-modal/finished-builds-filter-modal.component.scss
@@ -1,0 +1,11 @@
+.finished-build-jobs-filter-border-bottom {
+    border-bottom: 2px solid var(--overview-blue-border-color);
+}
+
+.finished-build-jobs-filter-background-accent {
+    background: var(--user-management-background-color);
+}
+
+.form-item-prepend {
+    width: 13%;
+}
